# openapi.json
{
...
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
...
    "/department/user": {
      "get": {
        "operationId": "getUserDepartment",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "Authorization",
            "required": false
          }
        ],
        "tags": [
          "department-controller"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}